chore: move calculateVotes to server utils
